+++ classes/class.ilPegasusHelperConfigGUI.php
@@ -8,18 +8,72 @@ require_once __DIR__ . '/../bootstrap.php';
 final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
 
     /**
+     * @var \ilPegasusHelperPlugin
+     */
+    var $pl;
+
+    static $ICON_CATEGORIES =  [
+        "course",
+        "folder",
+        "group",
+        "file",
+        "learningplace",
+        "learningmodule",
+        "link"
+    ];
+    static $ICON_WEB_DIR = "pegasushelper/theme/icons/";
+
+    public function __construct() {
+        $this->pl = new ilPegasusHelperPlugin();
+    }
+
+    /**
+     * copies the default icons to the directory for the synchronization with the app
+     */
+    static function copyDefaultIcons() {
+        global $DIC;
+
+        $customizingSourceDir = "global/plugins/Services/UIComponent/UserInterfaceHook/PegasusHelper/templates/icons/";
+        foreach(ilPegasusHelperConfigGUI::$ICON_CATEGORIES as $category) {
+            $fileName = "icon_$category.svg";
+            $fsWeb = $DIC->filesystem()->web();
+            $fsCustomizing = $DIC->filesystem()->customizing();
+            if($fsWeb->has(ilPegasusHelperConfigGUI::$ICON_WEB_DIR . $fileName)) {
+                $fsWeb->delete(ilPegasusHelperConfigGUI::$ICON_WEB_DIR . $fileName);
+            }
+            $fileStream = $fsCustomizing->readStream($customizingSourceDir . $fileName);
+            $fsWeb->writeStream(ilPegasusHelperConfigGUI::$ICON_WEB_DIR . $fileName, $fileStream);
+        }
+    }
+
+    /**
      * invoked by parent
      * @param $cmd string
      */
     public function performCommand($cmd) {
         global $ilTabs, $ilCtrl, $tpl;
-        $ilTabs->addSubTab("id_general", "General", $ilCtrl->getLinkTarget($this, "configure"));
-        $ilTabs->addSubTab("id_testing", "Testing", $ilCtrl->getLinkTarget($this, "testing"));
+        $ilTabs->addSubTab(
+            "id_general",
+            $this->pl->txt("tab_general"),
+            $ilCtrl->getLinkTarget($this, "general")
+        );
+        $ilTabs->addSubTab(
+            "id_theme",
+            $this->pl->txt("tab_app_theme"),
+            $ilCtrl->getLinkTarget($this, "theme")
+        );
+        $ilTabs->addSubTab(
+            "id_statistics",
+            $this->pl->txt("tab_statistics"),
+            $ilCtrl->getLinkTarget($this, "statistics")
+        );
+        $ilTabs->addSubTab(
+            "id_testing",
+            $this->pl->txt("tab_testing"),
+            $ilCtrl->getLinkTarget($this, "testing")
+        );
 
         switch ($cmd) {
-            case "saveColor":
-                $this->saveColor();
-                break;
             case "testing":
                 $ilTabs->setSubTabActive("id_testing");
                 $tpl->setContent($this->getTestsTableHtml());
@@ -28,10 +82,30 @@ final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
                 $ilTabs->setSubTabActive("id_testing");
                 $tpl->setContent($this->getTestsTableHtml(true));
                 break;
-            case "configure":
+            case "theme":
+                $ilTabs->setSubTabActive("id_theme");
+                $tpl->setContent($this->getColorFormHtml() . $this->getIconsForm()->getHTML());
+                break;
+            case "statistics":
+                $ilTabs->setSubTabActive("id_statistics");
+                $tpl->setContent($this->getTokenStatisticsHtml());
+                break;
+            case "theme_save_colors":
+                $this->saveColors();
+                break;
+            case "theme_reset_colors":
+                $this->resetColors();
+                break;
+            case "theme_save_icons":
+                $this->saveIcons();
+                break;
+            case "theme_reset_icons":
+                $this->resetIcons();
+                break;
+            case "general":
             default:
                 $ilTabs->setSubTabActive("id_general");
-                $tpl->setContent($this->getApiSecretFormHtml() . $this->getColorFormHtml());
+                $tpl->setContent($this->getApiSecretFormHtml());
                 break;
         }
     }
@@ -51,7 +125,7 @@ final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
                 $api_secret = $rec['api_secret'];
             }
 
-            $formApiUser->setTitle('ILIAS Pegasus API User');
+            $formApiUser->setTitle($this->pl->txt("form_api_secret"));
             $gui = new ilNonEditableValueGUI($api_key);
             $gui->setValue($api_secret);
             $formApiUser->addItem($gui);
@@ -61,11 +135,46 @@ final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
     }
 
     /**
-     * html of form for dynamic coloring
+     * html with token statistics
+     * @return string
+     */
+    protected function getTokenStatisticsHtml() {
+        $formTokensStatistics = new ilPropertyFormGUI();
+        $formTokensStatistics->setTitle($this->pl->txt("form_token_statistics"));
+
+        // get number of accesses for different durations
+        $differences = [
+            $this->pl->txt("txt_month") => 30,
+            $this->pl->txt("txt_quarter") => 90,
+            $this->pl->txt("txt_semester") => 180
+        ];
+
+        foreach($differences as $label => $dd) {
+            global $ilDB;
+            $sql = "SELECT COUNT(*) FROM ui_uihk_rest_refresh WHERE datediff(NOW(), created) < $dd";
+            $set = $ilDB->query($sql);
+            $counts = current($ilDB->fetchAssoc($set));
+
+            $gui = new ilNonEditableValueGUI($label);
+            $gui->setValue($counts);
+            $formTokensStatistics->addItem($gui);
+        }
+
+        return $formTokensStatistics->getHTML();
+    }
+
+    /**
+     * form for dynamic coloring
      * @return string
      */
     protected function getColorFormHtml() {
         global $ilDB, $ilCtrl;
+
+        $form = new ilPropertyFormGUI();
+        $form->setTitle($this->pl->txt("form_coloring"));
+        $form->setFormAction($ilCtrl->getFormAction($this));
+        $form->addCommandButton("theme_reset_colors", $this->pl->txt("button_reset"));
+        $form->addCommandButton("theme_save_colors", $this->pl->txt("button_save"));
 
         $primaryColor = "04427e";
         $contrastColor = 1;
@@ -78,46 +187,74 @@ final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
             }
         }
 
-
-        $formColor = new ilPropertyFormGUI();
-        $formColor->setTitle("App Theme Coloring");
-        $formColor->setFormAction($ilCtrl->getFormAction($this));
         // preview
         $thisDir = "Customizing/global/plugins/Services/UIComponent/UserInterfaceHook/PegasusHelper/classes";
         $tpl = new ilTemplate("tpl.theme_example.html", true, true, $thisDir);
         $tpl->setVariable('EX_TEXT', "&nbsp;ILIAS Pegasus&nbsp;");
         $tpl->setVariable('COLOR_PRIMARY', $primaryColor);
         $tpl->setVariable('COLOR_CONTRAST', $contrastColor ? "ffffff" : "000000");
-        $themeExample = new ilNonEditableValueGUI("Preview", "", true);
-        $themeExample->setInfo("The current theme");
+        $themeExample = new ilNonEditableValueGUI($this->pl->txt("txt_preview"), "", true);
+        $themeExample->setInfo($this->pl->txt("txt_info_preview"));
         $themeExample->setValue($tpl->get());
-        $formColor->addItem($themeExample);
+        $form->addItem($themeExample);
 
         // primary color
         require_once("./Services/Form/classes/class.ilColorPickerInputGUI.php");
-        $primaryInput = new ilColorPickerInputGUI("Primary color", "primary_color");
-        $primaryInput->setInfo("The main color for the theme of the app");
+        $primaryInput = new ilColorPickerInputGUI($this->pl->txt("txt_primary_color"), "primary_color");
+        $primaryInput->setInfo($this->pl->txt("txt_info_primary_color"));
         $primaryInput->setValue($primaryColor);
-        $formColor->addItem($primaryInput);
+        $form->addItem($primaryInput);
 
         // contrast color
         require_once("./Services/Form/classes/class.ilRadioGroupInputGUI.php");
         require_once("./Services/Form/classes/class.ilRadioOption.php");
-        $contrastInput = new ilRadioGroupInputGUI("Contrast color", "contrast_color");
-        $contrastInput->setInfo("The color of text in the app, which should be chosen dependent on the primary color");
-        $contrastInput->addOption(new ilRadioOption("White", 1));
-        $contrastInput->addOption(new ilRadioOption("Black", 0));
+        $contrastInput = new ilRadioGroupInputGUI($this->pl->txt("txt_contrast"), "contrast_color");
+        $contrastInput->setInfo($this->pl->txt("txt_info_contrast"));
+        $contrastInput->addOption(new ilRadioOption($this->pl->txt("button_white"), 1));
+        $contrastInput->addOption(new ilRadioOption($this->pl->txt("button_black"), 0));
         $contrastInput->setValue($contrastColor);
-        $formColor->addItem($contrastInput);
+        $form->addItem($contrastInput);
 
-        $formColor->addCommandButton("saveColor", "Save");
+        return $form->getHTML();
+    }
 
-        return $formColor->getHTML();
+    /**
+     * form for icons
+     * @return ilPropertyFormGUI
+     * @throws ilTemplateException
+     */
+    function getIconsForm() {
+        global $ilCtrl;
+
+        $form = new ilPropertyFormGUI();
+        $form->setTitle($this->pl->txt("form_icon"));
+        $form->setFormAction($ilCtrl->getFormAction($this));
+        $form->addCommandButton("theme_reset_icons", $this->pl->txt("button_reset"));
+        $form->addCommandButton("theme_save_icons", $this->pl->txt("button_save"));
+
+        foreach(ilPegasusHelperConfigGUI::$ICON_CATEGORIES as $category) {
+            // current item
+            $thisDir = "Customizing/global/plugins/Services/UIComponent/UserInterfaceHook/PegasusHelper/classes";
+            $clientName = CLIENT_ID;
+            $webDir = "data/$clientName/";
+            $tpl = new ilTemplate("tpl.icon.html", true, true, $thisDir);
+            $iconsDir = $webDir . ilPegasusHelperConfigGUI::$ICON_WEB_DIR;
+            $tpl->setVariable("SRC", $iconsDir . "icon_$category.svg");
+            $tpl->setVariable("SIZE", 45);
+            $icon = new ilNonEditableValueGUI(ucfirst($category), "", true);
+            $icon->setValue($tpl->get());
+            $form->addItem($icon);
+            // upload form
+            $fileUpload = new ilFileInputGUI("", "post_icon_$category");
+            $fileUpload->setSuffixes(["svg"]);
+            $form->addItem($fileUpload);
+        }
+
+        return $form;
     }
 
     /**
      * html of legend for tests
-     *
      * @param bool $external
      * @return string
      */
@@ -125,67 +262,70 @@ final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
         global $ilCtrl;
         include_once __DIR__ . "/class.ilPegasusTestingTableGUI.php";
         $tableLegend = new ilPegasusTestingTableGUI($this, "Status");
-        $tableLegend->setTitle("Legend for Tests");
+        $tableLegend->setTitle($this->pl->txt("tests_txt_legend"));
         require_once __DIR__ . "/class.ilPegasusTestingStatus.php";
         $dataLegend = [
             [
                 "status" => ilPegasusTestingStatus::T_STATUS_OK,
-                "test" => "test passed",
+                "test" => $this->pl->txt("tests_txt_passed"),
                 "info" => ""
             ],
             [
                 "status" => ilPegasusTestingStatus::T_STATUS_FAIL,
-                "test" => "test failed",
-                "info" => "a failed test indicates that the Pegasus App cannot operate correctly, the corresponding problem must be solved"
+                "test" => $this->pl->txt("tests_txt_failed"),
+                "info" => $this->pl->txt("tests_txt_failed_info")
             ],
             [
                 "status" => ilPegasusTestingStatus::T_STATUS_WARN,
-                "test" => "test resulted in a warning",
-                "info" => "depending on the setup of ILIAS, a test resulting in a warning may indicate a problem"
+                "test" => $this->pl->txt("tests_txt_warning"),
+                "info" => $this->pl->txt("tests_txt_warning_info")
             ],
             [
                 "status" => ilPegasusTestingStatus::T_STATUS_INCOMPLETE,
-                "test" => "it was not possible to run the test",
+                "test" => $this->pl->txt("tests_txt_aborted"),
                 "info" => ""
             ]
         ];
         $tableLegend->setData($dataLegend);
 
         // table with internal tests
-        $tableInt = new ilPegasusTestingTableGUI($this, "Name");
-        $tableInt->setTitle("Tests");
+        $tableInt = new ilPegasusTestingTableGUI($this, $this->pl->txt("tests_txt_name"));
+        $tableInt->setTitle($this->pl->txt("tests_txt_tests"));
         require_once __DIR__ . "/class.ilPegasusTesting.php";
         $tableInt->setData((new ilPegasusHelperTesting())->run("internal"));
 
         // table with external tests, run if $external is true
-        $tableExt = new ilPegasusTestingTableGUI($this, "Name");
-        $tableExt->setTitle("External Tests");
+        $tableExt = new ilPegasusTestingTableGUI($this, $this->pl->txt("tests_txt_name"));
+        $tableExt->setTitle($this->pl->txt("tests_txt_external_tests"));
         if($external) {
             require_once __DIR__ . "/class.ilPegasusTesting.php";
             $tableExt->setData((new ilPegasusHelperTesting())->run("external"));
         }
         $tableExt->setFormAction($ilCtrl->getFormAction($this));
-        $tableExt->addCommandButton("testing_run_external_tests", "Run external tests");
+        $tableExt->addCommandButton(
+            "testing_run_external_tests",
+            $this->pl->txt("tests_txt_run_external_tests")
+        );
 
         return $tableLegend->getHTML() . $tableInt->getHTML() . $tableExt->getHTML();
     }
 
     /**
-     * save input from the color form
+     * save input from the colors form
      */
-    protected function saveColor() {
+    protected function saveColors() {
         global $ilDB, $ilCtrl;
         $primaryColor = $_POST["primary_color"];
         $contrastColor = $_POST["contrast_color"];
 
         if(!preg_match("/^[0-9a-fA-F]{6}$/", $primaryColor)) {
-            ilUtil::sendFailure("App theme was not saved", true);
-            $ilCtrl->redirect($this, "configure");
+            ilUtil::sendFailure($this->pl->txt("msg_coloring_not_saved"), true);
+            $ilCtrl->redirect($this, "theme");
             return;
         }
 
         $values = array(
-            "primary_color" => array("text", $primaryColor),
+            "primary_color"  => array("text", $primaryColor),
             "contrast_color" => array("integer", $contrastColor)
         );
         $where = array(
@@ -193,8 +333,110 @@ final class ilPegasusHelperConfigGUI extends ilPluginConfigGUI {
         );
         $ilDB->update("ui_uihk_pegasus_theme", $values, $where);
 
-        ilUtil::sendSuccess("App theme saved successfully", true);
-        $ilCtrl->redirect($this, "configure");
+        $this->updateTimestamp();
+        ilUtil::sendSuccess($this->pl->txt("msg_coloring_saved"), true);
+        $ilCtrl->redirect($this, "theme");
+    }
+
+    /**
+     * reset colors to default values
+     */
+    protected function resetColors() {
+        global $ilDB, $ilCtrl;
+
+        $values = array(
+            "primary_color"  => array("text", "4a668b"),
+            "contrast_color" => array("integer", 1)
+        );
+        $where = array(
+            "id" => array("integer", 1)
+        );
+        $ilDB->update("ui_uihk_pegasus_theme", $values, $where);
+
+        $this->updateTimestamp();
+        ilUtil::sendSuccess($this->pl->txt("msg_coloring_reset"), true);
+        $ilCtrl->redirect($this, "theme");
+    }
+
+    protected function saveIcons() {
+        global $ilCtrl, $DIC;
+
+        $form = $this->getIconsForm();
+        if($form->checkInput()) {
+            // manage upload
+            if($DIC->upload()->hasBeenProcessed() !== true) {
+                if(PATH_TO_GHOSTSCRIPT !== "") {
+                    $DIC->upload()->register(new ilCountPDFPagesPreProcessors());
+                }
+            }
+            $DIC->upload()->process();
+
+            // for each category, set the new icon
+            $msgSuccess = "";
+            $msgFail = "";
+            foreach(ilPegasusHelperConfigGUI::$ICON_CATEGORIES as $category) {
+                $key = "post_icon_$category";
+                $file = $form->getInput($key);
+
+                foreach($DIC->upload()->getResults() as $result) {
+                    if($result->getStatus()->getCode() === \ILIAS\FileUpload\DTO\ProcessingStatus::OK) {
+                        $resultMatchesCategory = $file["tmp_name"] === $result->getPath();
+                        if($resultMatchesCategory) {
+                            $fileName = "icon_$category.svg";
+                            if($DIC->filesystem()->web()->has(ilPegasusHelperConfigGUI::$ICON_WEB_DIR . $fileName)) {
+                                $DIC->filesystem()->web()->delete(ilPegasusHelperConfigGUI::$ICON_WEB_DIR . $fileName);
+                            }
+                            $DIC->upload()->moveOneFileTo($result, ilPegasusHelperConfigGUI::$ICON_WEB_DIR, \ILIAS\FileUpload\Location::WEB, $fileName);
+                            $msgSuccess .= $this->pl->txt("msg_icon_saved_pre") . $category . $this->pl->txt("msg_icon_saved_post");
+                        }
+                    } else {
+                        if($result->getName()) {
+                            $msgFail = $this->pl->txt("msg_icon_not_uploaded");
+                        }
+                    }
+                }
+            }
+            $this->updateTimestamp();
+            // user feedback
+            if($msgSuccess) {
+                ilUtil::sendSuccess($msgSuccess, true);
+            }
+            if($msgFail) {
+                ilUtil::sendFailure($msgFail, true);
+            }
+        } else {
+            ilUtil::sendFailure($this->pl->txt("msg_icons_not_saved"), true);
+        }
+
+        $ilCtrl->redirect($this, "theme");
+    }
+
+    /**
+     * reset icons to default values
+     */
+    protected function resetIcons() {
+        global $ilCtrl;
+
+        ilPegasusHelperConfigGUI::copyDefaultIcons();
+
+        $this->updateTimestamp();
+        ilUtil::sendSuccess($this->pl->txt("msg_icons_reset"), true);
+        $ilCtrl->redirect($this, "theme");
+    }
+
+    /**
+     * sets the timestamp for the settings to the current time
+     */
+    private function updateTimestamp() {
+        global $ilDB;
+
+        $values = array(
+            "timestamp" => array("integer", time())
+        );
+        $where = array(
+            "id" => array("integer", 1)
+        );
+        $ilDB->update("ui_uihk_pegasus_theme", $values, $where);
     }
 }