+++ plugin.php
@@ -1,7 +1,7 @@
 <?php
 $id = 'sragpegasushelper';
-$version = '1.1.2';
+$version = '1.1.4';
 $ilias_min_version = '5.2.0';
-$ilias_max_version = '5.4.999';
-$responsible = 'Martin Studer, Nicolas Märchy, Nicolas Schäfli, Marc Schneiter, Jonathan Gerber';
+$ilias_max_version = '6.0';
+$responsible = 'Martin Studer, Nicolas Schäfli';
 $responsible_mail = 'info@studer-raimann.ch';