+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.4]
+### Added
+- added token statistics (number of logins for past 30 90 180 days)
+- configuration of REST client without requests to API
+
+## [1.1.3]
+### Added
+- app theme icons
+
+## [1.1.2]
+### Fixed
+- addition of table 'ui_uihk_pegasus_theme' 
+
 ## [1.1.1]
 ### Added
 - external testing